requirements.txt
Django==2.0
django-cors-headers==2.1.0
requests==2.18.4
django_celery==3.2.2
djangorestframework==3.8.2
pytz==2018.5
PyMySQL==0.8.0
